+++ include/kernel/page.h
@@ -50,7 +50,7 @@ void kfree(void *addr);
 void mm_init();
 void free_bootmem();
 size_t getfree();
-int heap_init(void *pool, void *start, void *end);
+size_t heap_init(void *pool, void *start, void *end);
 
 void *sys_brk(size_t size);
